Sprite Animation Basics
New Class Animation allows calculating of sprite frames based on a milisecond duration (to help me transition sprites form Asprite to the project). The animation class does not take any sprite data other than the animation frame range, a name for the animation (could probably replace with a integer for later enum animation manager), and a int list of milisecond delays (might need to swap to long if we get big enough animations)
Use Start animation to start it and updateFrame to calculate a new look based on real time passed.
setRepeat allows the animation to be repeated/looped